requirements/testing.txt
tox
tox-py
pytest>=6.1,<7.0
pytest-cov>=2.10.1,<3.0
pytest-django>=4.1.0,<5.0
importlib-metadata<5.0
